+++ picamera2/previews/q_gl_picamera2.py
@@ -111,24 +111,26 @@ class QGlPicamera2(QWidget):
         self.camera_notifier = QSocketNotifier(self.picamera2.notifyme_r,
                                                QSocketNotifier.Read, self)
         self.camera_notifier.activated.connect(self.handle_requests)
+        # Must always run cleanup when this widget goes away.
+        self.destroyed.connect(lambda: self.cleanup())
         self.running = True
 
     def cleanup(self):
-        with self.lock:
-            self.running = False
-            self.camera_notifier.deleteLater()
-            eglDestroySurface(self.egl.display, self.surface)
-            self.surface = None
-            # We may be hanging on to a request, return it to the camera system.
-            if self.current_request is not None and self.own_current:
-                self.current_request.release()
-            self.current_request = None
-            # We have to tell both the preview window and the Picamera2 object that we have
-            # disappeared.
-            if self.picamera2 is not None:
-                self.picamera2.detach_preview()
-            if self.preview_window is not None:
-                self.preview_window.qpicamera2 = None
+        if not self.running:
+            return
+        self.running = False
+        self.camera_notifier.deleteLater()
+        eglDestroySurface(self.egl.display, self.surface)
+        self.surface = None
+        # We may be hanging on to a request, return it to the camera system.
+        if self.current_request is not None and self.own_current:
+            self.current_request.release()
+        self.current_request = None
+        # We have to tell both the preview window and the Picamera2 object that we have
+        # disappeared.
+        self.picamera2.detach_preview()
+        if self.preview_window is not None:  # will be none when a proper Qt app
+            self.preview_window.qpicamera2 = None
 
     def closeEvent(self, event):
         self.cleanup()
@@ -300,6 +302,8 @@ class QGlPicamera2(QWidget):
     def set_overlay(self, overlay):
         if self.picamera2.camera_config is None:
             raise RuntimeError("Camera must be configured before setting overlay")
+        if self.picamera2.camera_config['buffer_count'] < 2:
+            raise RuntimeError("Need at least buffer_count=2 to set overlay")
 
         with self.lock:
             eglMakeCurrent(self.egl.display, self.surface, self.surface, self.egl.context)
@@ -381,6 +385,8 @@ class QGlPicamera2(QWidget):
 
     @pyqtSlot()
     def handle_requests(self):
+        if not self.running:
+            return
         self.picamera2.notifymeread.read()
         self.picamera2.process_requests(self)