+++ CHANGELOG.md
@@ -6,6 +6,30 @@
 
 ### Changed
 
+## 0.3.11 Beta Release 10
+
+### Added
+
+### Changed
+
+* The search path for tuning files (after the libcamera re-organisation) has been fixed.
+* The built-in configuration objects (picamera2.preview_configuration etc.) are updated to work better with raw streams now that the latest libcamera is more efficient when we use them.
+
+## 0.3.10 Beta Release 9
+
+### Added
+
+* Support for running multiple encoders, either on the same or different streams.
+* Pathlib objects now supported.
+
+### Changed
+
+* The camera can be stopped more quickly.
+* Fix problem passing modes from the sensor mode list to the configure methods.
+* H264 encoder quality was having no effect on bitrate - fixed.
+* ffmpegoutput - wait correctly for it to finish when we stop it.
+* Amount of CMA memory required is reduced for many use cases.
+
 ## 0.3.9 Beta Release 8
 
 ## Unreleased (on "next" branch)